requirements.txt
pydantic>=2.0.0
typing-extensions  
openai>=1.3.0
anthropic>=0.8.0
python-dotenv>=0.19.0
